control default latest version for avago and subnetevm (#2724)
* control default latest version for avago and subnetevm

* add rpc to json

* get latest cli version

* update json model

* update evm prompts

* update evm prompt

* update avalanchego latest

* add e2e test

* remove unncessary file

* lint

* try with lower versions

* update how we get newest avago version

* fix lint

* address comments

* fix ci

* fix ci

* fix lint

* lint

* move to dependencies pkg

* lint

---------
