+++ cmd/blockchaincmd/change_weight.go
@@ -363,7 +363,7 @@ func changeWeightACP99(
 	if err != nil {
 		return err
 	}
-	aggregatorLogger, err := signatureaggregator.NewSignatureAggregatorLoggerNewLogger(
+	aggregatorLogger, err := signatureaggregator.NewSignatureAggregatorLogger(
 		changeWeightFlags.SigAggFlags.AggregatorLogLevel,
 		changeWeightFlags.SigAggFlags.AggregatorLogToStdout,
 		app.GetAggregatorLogDir(clusterName),
