+++ pkg/models/compatibility.go
@@ -8,3 +8,13 @@ type VMCompatibility struct {
 }
 
 type AvagoCompatiblity map[string][]string
+
+type NetworkVersion struct {
+	LatestVersion string `json:"latest-version"`
+}
+
+type CLIDependencyMap struct {
+	RPC         int                       `json:"rpc"`
+	SubnetEVM   string                    `json:"subnet-evm"`
+	AvalancheGo map[string]NetworkVersion `json:"avalanchego"`
+}
