+++ pkg/node/local.go
@@ -9,6 +9,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/ava-labs/avalanche-cli/pkg/dependencies"
+
 	"github.com/ava-labs/avalanche-cli/pkg/application"
 	"github.com/ava-labs/avalanche-cli/pkg/constants"
 	"github.com/ava-labs/avalanche-cli/pkg/localnet"
@@ -26,13 +28,14 @@ import (
 func setupAvalancheGo(
 	app *application.Avalanche,
 	avalancheGoBinaryPath string,
-	avaGoVersionSetting AvalancheGoVersionSettings,
+	avaGoVersionSetting dependencies.AvalancheGoVersionSettings,
+	network models.Network,
 	printFunc func(msg string, args ...interface{}),
 ) (string, error) {
 	var err error
 	avalancheGoVersion := ""
 	if avalancheGoBinaryPath == "" {
-		avalancheGoVersion, err = GetAvalancheGoVersion(app, avaGoVersionSetting)
+		avalancheGoVersion, err = dependencies.GetAvalancheGoVersion(app, avaGoVersionSetting, network)
 		if err != nil {
 			return "", err
 		}
@@ -54,7 +57,7 @@ func StartLocalNode(
 	defaultFlags map[string]interface{},
 	connectionSettings localnet.ConnectionSettings,
 	nodeSettings []localnet.NodeSetting,
-	avaGoVersionSetting AvalancheGoVersionSettings,
+	avaGoVersionSetting dependencies.AvalancheGoVersionSettings,
 	network models.Network,
 ) error {
 	// initializes directories
@@ -78,6 +81,7 @@ func StartLocalNode(
 			app,
 			avalancheGoBinaryPath,
 			avaGoVersionSetting,
+			network,
 			ux.Logger.PrintToUser,
 		)
 		if err != nil {
