+++ versions/latest.json
@@ -1,0 +1,26 @@
+{
+  "subnet-evm": "v0.7.3",
+  "rpc": 39,
+  "avalanchego": {
+    "Local Network": {
+      "latest-version": "v1.13.0",
+      "require-prerelease": false,
+      "prerelease-version": ""
+    },
+    "DevNet": {
+      "latest-version": "v1.13.0",
+      "require-prerelease": false,
+      "prerelease-version": ""
+    },
+    "Fuji": {
+      "latest-version": "v1.13.0",
+      "require-prerelease": false,
+      "prerelease-version": ""
+    },
+    "Mainnet": {
+      "latest-version": "v1.13.0",
+      "require-prerelease": false,
+      "prerelease-version": ""
+    }
+  }
+}
